HATR nodes merged
